warehouse/requirements.txt
requests
flask-lambda-python36
Werkzeug==2.0.3
boto3